extract nav menu from nav components
