+++ web-app/src/app/(app)/components/app-header.tsx
@@ -6,12 +6,10 @@ import UserAvatar from "./user-avatar";
 export default function AppHeader() {
   return (
     <header className="border-grid sticky top-0 z-50 flex h-[64px] w-full items-center border-b bg-background/95 px-4 py-3">
-      <div className="flex w-full items-center">
-        <CustomTrigger when="invisible" />
-        <MainNav />
-        <MobileNav />
-        <UserAvatar />
-      </div>
+      <CustomTrigger when="invisible" />
+      <MainNav />
+      <MobileNav />
+      <UserAvatar />
     </header>
   );
 }
