+++ web-app/src/app/(app)/components/mobile-nav.tsx
@@ -1,9 +1,4 @@
-"use client";
-
 import { ArrowLeftFromLine } from "lucide-react";
-import Link from "next/link";
-import { usePathname } from "next/navigation";
-import * as React from "react";
 
 import { SokosumiLogo } from "@/components/masumi-logos";
 import { Button } from "@/components/ui/button";
@@ -15,23 +10,14 @@ import {
   SheetTitle,
   SheetTrigger,
 } from "@/components/ui/sheet";
-import { cn } from "@/lib/utils";
 
-const navItems = [
-  { href: "/dashboard", label: "Home" },
-  { href: "/jobs", label: "Jobs" },
-  { href: "/billing", label: "Billing" },
-  { href: "/settings", label: "Settings" },
-];
+import NavMenu from "./nav-menu";
 
 export default function MobileNav() {
-  const [open, setOpen] = React.useState(false);
-  const pathname = usePathname();
-
   return (
-    <Sheet open={open} onOpenChange={setOpen}>
+    <Sheet>
       <SheetTrigger asChild>
-        <div className="ml-2 mr-2 flex flex-1 justify-end md:hidden">
+        <div className="mx-2 flex flex-1 justify-end md:hidden">
           <Button variant="outline" size="icon">
             <ArrowLeftFromLine />
           </Button>
@@ -42,33 +28,11 @@ export default function MobileNav() {
           <SheetTitle className="flex justify-center">
             <SokosumiLogo />
           </SheetTitle>
-          <SheetDescription></SheetDescription>
+          <SheetDescription />
         </SheetHeader>
-        <ul className="mt-4 flex flex-col gap-y-2 overflow-y-auto">
-          {navItems.map((nav) => {
-            const isActive = pathname.startsWith(nav.href);
-            if (!nav.href) return;
-            return (
-              <li className="w-full p-2" key={nav.href}>
-                <Link
-                  href={nav.href}
-                  className={cn(
-                    "flex items-center gap-2 text-lg transition hover:text-foreground/80",
-                    {
-                      "text-foreground underline": isActive,
-                      "text-foreground/50": !isActive,
-                    },
-                  )}
-                >
-                  {nav.label}
-                </Link>
-              </li>
-            );
-          })}
-        </ul>
-        <div className="mx-2 my-4 text-sm font-bold text-muted-foreground">
-          Credits balance: 6901
-        </div>
+        <nav className="mt-4">
+          <NavMenu className="flex-col gap-4 text-base" />
+        </nav>
       </SheetContent>
     </Sheet>
   );
